requirements.txt
paddleocr==2.9.1
pandas==2.2.3
requests==2.32.3
slack_sdk==3.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.36